Refactoring all components that used StatusChip to use the new ones.
