+++ src/pages/public_registration/action_accordions/PublishingAccordion.tsx
@@ -8,8 +8,8 @@ import { useDispatch, useSelector } from 'react-redux';
 import { Link, useLocation } from 'react-router';
 import { useDuplicateRegistrationSearch } from '../../../api/hooks/useDuplicateRegistrationSearch';
 import { publishRegistration } from '../../../api/registrationApi';
+import { TicketStatusChip } from '../../../components/_molecules/status-chip/TicketStatusChip';
 import { RegistrationErrorActions } from '../../../components/RegistrationErrorActions';
-import { TicketStatusChip } from '../../../components/StatusChip';
 import { HorizontalBox } from '../../../components/styled/Wrappers';
 import { ActionPanelContext } from '../../../context/ActionPanelContext';
 import { LandingPageContext } from '../../../context/LandingPageContext';
